# apps/ebpco-api/contract/citizen-endpoints.openapi.yaml
...
openapi: 3.1.0
info:
  title: eBPCO citizen endpoints — permit, documents, resubmission
  version: 1.0.0
  description: |
    Shipped in eBPCOBackend: C-1 at f7eb40e, C-2 at 5a0f18a, D-8 at 18028a8.

    All three require a citizen bearer token. All three answer 404 for an
    application that is not the caller's, which is the SAME 404 as one that does
    not exist: a reference number that resolves differently would confirm that a
    neighbour has applied.
paths:
...
  /me:
    get:
      operationId: getProfile
      summary: The signed-in citizen's own record
      security: [{ bearerAuth: [profile:read] }]
      responses:
        '200':
          description: The account and, for a citizen, their profile.
          content: { application/json: { schema: { $ref: '#/components/schemas/CitizenProfile' } } }

    patch:
      operationId: rectifyProfile
      summary: Correct what the LGU holds about you (RA 10173 §16(d))
      description: |
        A statutory right, not a settings screen: the change is audited, and a
        field that cannot be applied is REFUSED rather than ignored.

        **`email` is deliberately absent and sending it is a 400.** It is the
        sign-in identity, so changing it is a transfer of who can reach the
        account rather than a correction; doing that without first proving
        control of the new address would make this an account-takeover route.
        The schema is `additionalProperties: false` so a client is told, instead
        of watching the field be dropped and believing the address changed.

        **At least one field is required.** An empty body is a 400: answering
        200 would report a correction that never happened.

        **Changing `mobileNumber` un-verifies it.** The verification belonged to
        the OLD number -- a code was sent there and answered from there -- so
        `mobileVerifiedAt` is cleared and any pending challenge for the channel
        is deleted, or a code sent to the old phone could be confirmed
        afterwards and would verify the new number. The response says
        `mobileVerificationCleared: true` so a client re-prompts rather than
        leaving a citizen with an unverified contact they believe is verified.

        This applies the change immediately rather than holding it pending
        verification, and that is forced by the deployment rather than
        preferred: `POST /me/contacts/mobile/request` currently answers
        `delivery: "not-sent"` because the LGU has no message provider
        configured, so under a pending model a number change could never
        complete and every citizen who moved would be stranded on the old
        number.

        **Staff cannot use this.** An officer's name is set by the office on
        their access request and stands against their acts in the audit trail;
        correcting it is an administrator's job. A staff token gets 422.
      security: [{ bearerAuth: [profile:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileCorrection' }
      responses:
        '200':
          description: The corrected record, read back.
          content: { application/json: { schema: { $ref: '#/components/schemas/CorrectedProfile' } } }
        '400':
          description: |
            An unknown field (`email` among them), an empty body, a mobile
            number that is not `09XXXXXXXXX` or `+639XXXXXXXXX`, a `postalCode`
            that is not four digits, or an EMPTY STRING where a clear was meant
            -- see `ProfileCorrection`.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
        '422':
          description: This account has no applicant profile to correct (a staff account).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  /applications/{applicationId}/documents:
    get:
      operationId: listApplicationDocuments
      summary: Every document on the application, and the office's verdict on each
      description: |
        A BARE ARRAY, not an envelope — the same shape as the timeline. Empty
        array for an application of theirs with nothing uploaded, which is a
        different answer from 404 and a true one.

        This is the only route that returns a document id, which content
        retrieval and resubmission both require.
      security: [{ bearerAuth: [applications:read] }]
      parameters:
        - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The documents, oldest upload first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApplicationDocument' }
        '404':
          description: Not the caller's application, or no such application.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ProfileCorrection:
      type: object
      additionalProperties: false
      minProperties: 1
      description: |
        Every field optional; at least one required.

        **ABSENT and NULL mean different things.** Absent leaves a field as it
        is; `null` CLEARS it. A citizen who typed a middle name by mistake, or
        has none, must be able to remove it -- a right to correct that cannot
        remove is half a right.

        **An empty string is NOT a clear and is refused (400).** Every field is
        `minLength: 1` when a string is supplied, so `""` fails validation
        rather than quietly erasing the value. Send `null`. This is the detail
        the citizen web lane measured off the source and reported as
        unguessable, which is why it is stated here.
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        middleName:
          type: [string, 'null']
          minLength: 1
          maxLength: 100
          description: Send `null` to clear. `""` is a 400, not a clear.
        lastName: { type: string, minLength: 1, maxLength: 100 }
        mobileNumber:
          type: string
          pattern: '^(09\\d{9}|\\+639\\d{9})$'
          description: |
            `09XXXXXXXXX` or `+639XXXXXXXXX`. **A bare `639XXXXXXXXX` is
            REFUSED** -- reject it client-side rather than spending a round trip
            on a 400. Changing this un-verifies the number.
        street:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: |
            `street`, NOT `address`. `businesses` already calls it street, and
            two spellings of one idea in one service is what D-10 spent a
            migration undoing.
        barangay: { type: [string, 'null'], minLength: 1, maxLength: 100 }
        city: { type: [string, 'null'], minLength: 1, maxLength: 100 }
        province: { type: [string, 'null'], minLength: 1, maxLength: 100 }
        postalCode:
          type: [string, 'null']
          pattern: '^[0-9]{4}$'
          description: Four digits. An address the office cannot post to is worse than none.
    CitizenProfile:
      type: object
      description: |
        `GET /me`. The address parts are null for every citizen who has not been
        asked -- **null means NOT RECORDED**, never a blank the citizen chose to
        leave empty.
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [applicant, staff] }
        email: { type: string, format: email }
        emailVerifiedAt: { type: [string, 'null'], format: date-time }
        firstName: { type: [string, 'null'] }
        middleName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        mobileNumber: { type: [string, 'null'] }
        street: { type: [string, 'null'] }
        barangay: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        province: { type: [string, 'null'] }
        postalCode: { type: [string, 'null'] }
        dateOfBirth: { type: [string, 'null'] }
        sex: { type: [string, 'null'], enum: [Male, Female, 'Prefer not to say', null] }
        civilStatus: { type: [string, 'null'], enum: [Single, Married, Widowed, Separated, Divorced, null] }
        nationality: { type: [string, 'null'] }
        mobileVerifiedAt:
          type: [string, 'null']
          format: date-time
          description: |
            On the READ as well as the correction. Verification is a property of
            the account, and a client needs it on load rather than only after a
            change — its absence here is what made one shared interface lie.
    CorrectedProfile:
      type: object
      additionalProperties: false
      description: |
        A STRICT SUPERSET of `CitizenProfile`, plus one field.

        Stated because it was not always true, and the difference cost a lane
        real work: `GET /me` and `PATCH /me` once overlapped without either
        containing the other — the read carried `id`/`kind`/`email`, the
        correction carried the verification fields — so a client typing ONE
        interface for both declared something absent at runtime whichever way it
        chose. `mobileVerifiedAt` was the casualty: `undefined` renders as a
        verified-looking blank on the one screen where verification state is the
        thing being shown.

        `mobileVerificationCleared` is the only field here a read cannot answer,
        because it describes what THIS request did rather than what the account
        is. Everything else is the profile.
      required:
        [id, kind, email, emailVerifiedAt, firstName, middleName, lastName,
         mobileNumber, street, barangay, city, province, postalCode,
         dateOfBirth, sex, civilStatus, nationality,
         mobileVerifiedAt, mobileVerificationCleared]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [applicant, staff] }
        email: { type: [string, 'null'], format: email }
        emailVerifiedAt: { type: [string, 'null'], format: date-time }
        firstName: { type: [string, 'null'] }
        middleName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        mobileNumber: { type: [string, 'null'] }
        street: { type: [string, 'null'] }
        barangay: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        province: { type: [string, 'null'] }
        postalCode: { type: [string, 'null'] }
        dateOfBirth:
          type: [string, 'null']
          description: 'YYYY-MM-DD, or null when not recorded (migration 038 — nullable for every applicant that registered before this field existed).'
        sex: { type: [string, 'null'], enum: [Male, Female, 'Prefer not to say', null] }
        civilStatus: { type: [string, 'null'], enum: [Single, Married, Widowed, Separated, Divorced, null] }
        nationality: { type: [string, 'null'] }
        mobileVerifiedAt: { type: [string, 'null'], format: date-time }
        mobileVerificationCleared:
          type: boolean
          description: |
            True when this request changed the number and therefore cleared its
            verification. Read it and re-prompt: a citizen left believing an
            unverified contact is verified will not receive what the office
            sends there.
...
    ApplicationDocument:
      type: object
      additionalProperties: false
      required:
        [id, label, fileName, contentType, byteSize, sha256, uploadedAt, expiresOn,
         certifiedOn, requirementCode, reviewStatus, reviewedAt, reviewReason,
         reviewRemark, supersedesDocumentId, supersededByDocumentId, scanCleared,
         quarantined]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        fileName: { type: string }
        contentType: { type: string }
        byteSize:
          type: string
          description: A STRING, not a number — the column is a bigint and JSON numbers lose precision.
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        uploadedAt: { type: string, format: date-time }
        expiresOn: { type: [string, 'null'], format: date }
        certifiedOn:
          type: [string, 'null']
          format: date
          description: |
            When the ISSUING OFFICE certified this document, read off its face.

            Required by the Municipal ruling of 2026-09-03: the admin note on a
            reused document says it is reused **and the date it was certified**.
            An expiry cannot answer that, and an upload date is when the file
            reached this service — three different dates, and none may stand in
            for another. A test asserts they are distinct.

            **NULL MEANS NOT RECORDED**, never that the document was
            uncertified. Nobody supplies it yet — who does is a workflow
            decision with the Municipality — so it is null on every document
            today. Render "certification date not recorded"; do NOT fall back to
            the upload date, which would tell an officer a document was
            certified on the day it reached us.
          examples: ['2024-03-12']
        requirementCode:
          type: [string, 'null']
          description: |
            Which checklist entry this document answers. NULL MEANS NOT
            ATTRIBUTED — nobody recorded which requirement it is for — and never
            that it answers none. Every document uploaded before this existed
            carries null.
        reviewStatus:
          type: [string, 'null']
          enum: ['Missing', 'Uploaded', 'Submitted', 'Under Review', 'Accepted',
                 'Rejected', 'Revision Required', 'Expired', null]
          description: |
            The OFFICER's verdict. NULL MEANS NOBODY HAS LOOKED YET — it does not
            mean nothing is wrong. Rendering null as a tick would tell an
            applicant their document passed when it has not been opened.
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewReason:
          description: |
            The standard, reusable reason. Null when none was cited. USE `label`
            for display rather than keeping your own copy of the catalogue: the
            LGU can edit it, and a client-side copy drifts. `code` is what to
            switch on.
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [code, label, description]
              properties:
                code: { type: string, examples: ['illegible'] }
                label: { type: string, examples: ['Illegible'] }
                description: { type: string }
        reviewRemark:
          type: [string, 'null']
          description: Free text written for THIS applicant about THIS document.
        supersedesDocumentId: { type: [string, 'null'], format: uuid }
        supersededByDocumentId: { type: [string, 'null'], format: uuid }
        scanCleared:
          type: boolean
          description: Malware scan completed and passed. A DIFFERENT axis from reviewStatus.
        quarantined:
          type: boolean
          description: |
            Held by the malware scanner. Not an evaluation outcome — an officer's
            rejection does not mean a virus, and a quarantined file is not a
            verdict on the application.
...
    Problem:
      type: object
      description: RFC 9457. `detail` is the human-readable half.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }